tests(e2e): rename event schemas file to field-contracts
